requirements.txt
asgiref==3.9.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
fonttools==4.59.0
imageio==2.37.0
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
PassportEye==2.2.2
pdfminer==20191125
pillow==11.3.0
pycountry==24.6.1
pycryptodome==3.23.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tifffile==2025.6.11
tzdata==2025.2
